Bring back the *mandatory* VAYRA_SETTINGS.
Not only do I not like this, but it's also not using optParam() but rather getParam()
so it's going to be crashing alot due to JSONExceptions. Furthermore, the file itself
fails the JSON lint check due to some random comments.

This whole piece needs to be revisited and cleaned up. Starting with the mod's JAR side first.