Get rid of the 'packed' folder
While the packed/FINAL was a decent idea, it turns out that any tool other than Notepad++
either breaks CSVs or somehow screws up something. So, while it was good for a general overview of
"what else is there to pull in" it proved much better and time-saving to just pull directly to
src/data instead, and commit after verifying.

tl;dr the automerge sucked, theres no need for this folder to exist anymore - it's done it's purpose.