Command works fine with a list of queries, invalid token issue should also be fixed
Or in other words, commands such as
'findships { cargo > 500,crew > 800,ballistic > 3 }'
'findships large ballistic > 5'
'findships { universal > 2,frigate }'
all work fine now.

However, commands which involved invalid tokens
'findships { cargo > 400,ballistic > 3,flightbay > 2 }'
'findships beds > 4'
were causing a nullpointer somewhere. Due to the lack of a debugger and debugging facilities,
might just be easier to verify if the first token in each expression is valid by checking
if it's a weapon/size/ship parameter than to rely on not making any mistakes anywhere.

Other than that, it works pretty solid, the results all check out.

One tidbit though: some results seem kinda duplicated.
for instance, for the following command "findships large ballistic > 6" the top two results
are IndEvo_brimir and IndEvo_brimir_default_D

I'm guessing one is a variant and the other is the hull, because addship IndEvo_brimir_default_D doesn't work
but it works fine for the other one. So something else needs to be cut out of the set that is queried.

However, since I didn't commit in a long while, lets keep that for a later commit.