UtilityConsole FindShips command (#1)
* Initial commit ready for testing; doesn't seem like it's working yet

* Fixed the mod, got some good results back.

Alright, now that it seems to be functioning, there's a few more things to do:
1) proper formatting of the output. Due to name length differences, the results are scattered. Let's stringformat the shit out of that.
2) make it work for a list of expressions
3) more testing of edgecases and broken inputs

* string formatting; extract stringifying results into a new method

* Command works fine with a list of queries, invalid token issue should also be fixed

Or in other words, commands such as
'findships { cargo > 500,crew > 800,ballistic > 3 }'
'findships large ballistic > 5'
'findships { universal > 2,frigate }'
all work fine now.

However, commands which involved invalid tokens
'findships { cargo > 400,ballistic > 3,flightbay > 2 }'
'findships beds > 4'
were causing a nullpointer somewhere. Due to the lack of a debugger and debugging facilities,
might just be easier to verify if the first token in each expression is valid by checking
if it's a weapon/size/ship parameter than to rely on not making any mistakes anywhere.

Other than that, it works pretty solid, the results all check out.

One tidbit though: some results seem kinda duplicated.
for instance, for the following command "findships large ballistic > 6" the top two results
are IndEvo_brimir and IndEvo_brimir_default_D

I'm guessing one is a variant and the other is the hull, because addship IndEvo_brimir_default_D doesn't work
but it works fine for the other one. So something else needs to be cut out of the set that is queried.

However, since I didn't commit in a long while, lets keep that for a later commit.

* Fixed all issues - results are still kinda doubled though

Everything seems to be working great. Even the spaces are no longer that much of an issue.

Few things to look at:
 - duplicated results
 - string.format on listships

* Added help in failure scenarios for findships

* Applied formatting to ListShips

* Fix string formatting for RefillCR command

* Latest JAR

* Address CR comments