Added support for ordnance points to FindShips
I was interested in which was the beefiest ship. And it turns out that there's quite a few
1000 OP weird-looking frigate ships that might be plug'n'play modules to other ships or i dunno.

But there are ships with 500 OP.

Also, this thing works from first try.

Added necessary changes to FindShips script (it's input verification part and remapping to enum),
the commands.csv to mention it's ok to use it, and ParameterCriterion (mostly matches() and countMatches() methods)

Also added the new JAR too.