More refactoring; fully covered Bulk Transport skill
As the method grew, it became apparent that parts of it should be extracted to
other private methods. So the whole LunaListener was broken into methods
that handle each of the present tabs, while the Skills tab related method
was further broken down to cover Combat, Leadership, Technology and Industry skills.

Furthermore, each of those skill group methods are broken into calling into each skill-related
method.

This setup will allow for maximum clarity and minimum boilerplate... while the ModPlugin class
will grow uncontrollably. Even then, these methods *could* just be extracted to different classes,
and have the LunaListener call into them instead of all of them being in the same file.

But that's a later problem.