WIP commit - ran into some technical problems.
Thing is, using reflection is blocked.
But using reflection to access reflection is allowed.

So the whole approach of how to strip members of their 'final' modifier needs to be
redone from scratch, using the reflection way of calling reflection...

This is a commit before i mess something else up.