another WIP commit; replaced reflection with MethodHandles; still no dice
Now that the original reflection-based approach has been replaced with an alternative
MethodHandles way of using reflection, i still have no idea why it doesn't work.

The ReflectionUtils::modifyFinalField() might not even be entered due to lack of log
at the start of the method, and similar applies to ReflectionUtils::setOrUnsetFinal() method
as well.

So, saving this, for just in case, so it's not lost before i refactor this a bit more.