Nuke VayraTags
The public message is that it's deprecated and no longer needed.

The real hidden message is that it was the crashcode-wannabe, not really crashing your save
but it had a dumbass "obfuscated" list of factions that it had a love/hate relationship going on.
In some cases, it would quadruple the chances of good stuff happening to them, in other (usual) case
it just made them (or their whole owning system) forever-enemies with the player.

The mods in question are: new_galactic_order, diableavionics, sindrian_diktat, aria, fpe, legioinfernalis
It also used to look at a particular planet - aesir, and do stupid shit if it was found.

Well, no more. Real petty move Vayra and it's bad UX.