Resolve issues #18 #19 #20 - ConcurrentModificationException
Ok, these three issues were all duplicates of the same issue.
On the other hand, the way this code was written, this crash would
happen 100% of the time, and is a legitimate textbook case of a
ConcurrentModificationException, where items were being removed
while the list was being iterated over.

Switching over to an iterator and removing via iterator should
fix the issue.
As reported by the immediate beta-tester - it resolved the issue.