Add two mining ships/weapon CSV files suggested by PNG1
As mentioned on discord, turns out this feature is available in some other
Vayra Ships version but not the one contained in the VayraMerged mod.

Quoting:
> Hi. Not sure it's really a priority at current or anything, but I noticed when I was playing earlier today that the various nex mining values for Vayra ships and weapons didn't make it into the combined modpack
pulled these from the most recent version of the vayra ship pack I could find with a trivial amount of searching

So, why not. Screenshots apparently prove that it works so put them in.