Fix issue 31 - IndexOutOfBounds in some D-mod in vayramerged
Issue ref:
https://github.com/RkShaRkz/Starsector-IntelliJ-Template/issues/31

The obvious root cause of the issue is that we're iterating over the size of one list, but we're changing three lists
using those indices.

Obviously, this means that the lists aren't always of the same size, but if we add a size check before changing them
that should fix the issue. Crap solution, but should fix it.

Added size checks before each of those things are done.