Fix issue 31 (#32)
* Fix issue 31 - IndexOutOfBounds in some D-mod in vayramerged

Issue ref:
https://github.com/RkShaRkz/Starsector-IntelliJ-Template/issues/31

The obvious root cause of the issue is that we're iterating over the size of one list, but we're changing three lists
using those indices.

Obviously, this means that the lists aren't always of the same size, but if we add a size check before changing them
that should fix the issue. Crap solution, but should fix it.

Added size checks before each of those things are done.

* Refactor and improve the previous solution

While the previous solution was somewhat alright, in a sense that it was using the same 'move' value
for all angle offsets, it changed the original implementation of using the same move value for
the HardpointAngleOffset, HiddenAngleOffset and TurretAngleOffset tuple.

So the new solution is to generate a 'move' array for each weapon, and let each whatever-AngleOffset pick
a value from the array. This will lead to HardpointAngleOffset[0], HiddenAngleOffset[0] and TurretAngleOffset[0]
all using the identical 'move' value, and so on for all of their subsequent indices.

Since the generated array is the max() of all three sizes, lists that are shorter than that simply won't use the extra
value. I think this is better than the previous one.