Alternative fix for issue #20
Well, since I'm using an iterator, and the ConcurrentModificationException is
maybe/obviously still happening - there's only one explanation for that.

Stuff is being added while the list is being iterated over and "fucked up participants"
are being removed.

So, instead of grabbing the iterator of the getSharedData().getParticipatingFactions() list,
lets first make a copy of that list, and then iterate over that copy.
Hopefully that makes it go away for good.