Yet another fix for issue #20
In case the ConcurrentModificationException is still happening somehow, which it should not,
I don't get where it's coming from because this file is ran in a single thread and there should
be no way for all these procedurally-later things to happen at the same time.

But, change things up a bit.