Hotfix for issue 39: IllegalArgumentException in VayraScreeningFormation::getStatusData()
Well, the method is just a big switch (when) and as usual, it default to throwing an
IllegalArgumentException when it encounters an index it did nothing for.

So instead of trying to use 2 indices to cover all 7 bonuses, the when() was extended with
five more cases to cover the bonuses/penalties which were lumped together into one of the two
original ones.

Tried reproing, crash no longer reproed; I didn't find where this text was even visible so...
couldn't quite validate whether it even shows up. But if it does - it's hidden somewhere.