Address IndexOutOfBounds problems in VayraJoachimEffect
This is the identical issue as in VayraDamagedOptics. However, on second look, it seems
that I have restructured the code - instead of rotating all 3 things in sync, I'll now
rotate them in order, which is probably *not* what should be happening...

But the OOB access is fixed for sure now.