Add LunaSettings and the new flag to VayraColonialManager
The settings should contain only one thing so far, but the whole 'architecture' was
setup alright to allow for future extensibility.

VayraColonialManager also got the new UPGRADES_DISABLED flag too, defaulting to 'false'
both in the java code, and the LunaSettings.csv

EDIT: added a semicolon where it was missing
EDIT2: fixed the icon in LunaSettings too and everything is hooked up properly.