Severe refactor - and it still works :)
So, before each method used to look for and update the stored maps and maybe resave them as well
Now, we only find the generator once (per module / fleetmember) and search for the emitters once;
once they are found, they are cached in the CustomData to be retrieved later on.

One small, new caveat is that the advanceInCombat() method no longer does anything if it can't
find the generator - now it immediatelly returns there.

However, the old way of "each method searches for generator/emitters" is now extracted and done only once.

I don't care about doing that for the projectiles and jitters. Just need to test caliph vs caliph and see
how their shield fluxes behave (ie - will they be shared or not)