Fix the caliph shield (and the vayra_modular_shields hullmod too) (#55)
* Fix the caliph shield (and the vayra_modular_shields hullmod too)

The problem with caliph's shields was that it was overloading from just one hit from a frigate.
The initial bug report said it's going to overload on it's own from nothing, however i wasn't able to repro that.

Because of the weird getCustomData() approach that **never saved the maps it was getting or generating**
the ship just went into overload super-fast. (tl;dr it was never draining any flux)

Now that the maps *are* being saved, it doesn't overload as fast at all, however there might still be
hidden caveats - namely:
- can two caliphs fight against each other and not collide? (not have their maps collide)
- do these things saved under a ShipAPI key get removed at all, at any time? these keys are surely not valid forever.
- does this getCustomData() super-map keep on growing or not?

These things should all be explored and accounted for, to avoid "yea we fixed the shields but caused huge amounts of memory leaks"
situation.

Also, the logs need to be muted or wrapped in the VAYRA_DEBUG call, but i'd use something similar to GuardianShield's
logFlux() method instead.

* Severe refactor - and it still works :)

So, before each method used to look for and update the stored maps and maybe resave them as well
Now, we only find the generator once (per module / fleetmember) and search for the emitters once;
once they are found, they are cached in the CustomData to be retrieved later on.

One small, new caveat is that the advanceInCombat() method no longer does anything if it can't
find the generator - now it immediatelly returns there.

However, the old way of "each method searches for generator/emitters" is now extracted and done only once.

I don't care about doing that for the projectiles and jitters. Just need to test caliph vs caliph and see
how their shield fluxes behave (ie - will they be shared or not)

* Mute logs + new JAR

Oh, and caliph vs caliph works fine :)

* Address CR comments and new JAR