Fix VayraPersonBountyManager rare bounty flagship CSV parsing
Well well well Vayra, yet again you suck.

The short story is - the weight of all rare bounty flagships is a number,
which was expected to be a double to be casted into float.

And of course, since we read the merged CSV on a row-by-row basis,
the weight was simply being passed in as 'row.getDouble("weight")'
which was *always* causing a JSONException("weight" is not a number).

And of course it's not a number, it's a string in a CSV file and has to be parsed-in
as such.

So this commit reworks the approach to read the "weight" as a string, then parse
it into the double.

Additionally, since this thing obviously never worked and was never noticed that it'd
read alot of empty lines into RareBountyFlagshipData classes, I made the 'variantId'
the deciding factor in whether a valid rare bounty was read or not; if it's empty or null
we don't create the RareBountyFlagshipData and add it to the RARE_FLAGSHIP list.

Also shame on me for not noticing this sooner but I didn't have that many crash-on-start problems
like I did recently.