Buffed LunaSettings CSV with dividers and extra lines for readability
The D-Mod "sections" were divided with both a header that mentions their in-game and code name
as well as empty lines full of commas to more easily distinguish between sections.

Furthermore, the 'dmod name' were put in brackets to paint them yellow in the Luna Settings menu
so they pop more and are more easily distinguished among the rest of the identical text.

UPDATE: made L'Interstellaire pop as well because why not.
UPDATE2: turns out, brackets don't work in settings' names, only their descriptions. so, removed brackets from names.
UPDATE3: there was a 'for for' typo in every 'enemy' setting's description... fixed it now.