Make VayraDamagedEverything / Performance Irregularities work with Rugged Construction
Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

Also turned it up a notch, 1% malfunction seemed a bit too low now that it's fixed up and parametrizable
from luna settings. 5% (+scaling) for both weapons and engine should do fine.

But CR remained at 10% just because "you can turn it off from the settings if you don't like it, shush"