Make VayraDamagedFighterWeapons / Malformed Fighter Weapons work with Rugged Construction
Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

Split up the recoil malus from the accuracy malus.
Decrease accuracy penalty but increase recoil penalty.

Finally, update the string in hull_mods.csv so that the recoil penalty is mentioned.