Make VayraDamagedMissiles / Flawed Missile Racks work with Rugged Construction
Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

Also, somewhat change up what the dmod does; instead of just bringing the ammo count
and RoF down, now it instead brings the ammo count and guidance down but brings RoF up!

Should be somewhat more fun to use I guess and not that much of a bore to get.