Make VayraDamagedTurrets / Misshapen Turret Gyros work with Rugged Construction
Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

Changed the dmod up a bit, instead of giving 25% turnrate and health penalties,
it now gives 15% turnrate and health penalties, and a new 5% chance of weapon malfunction.