Make VayraLessIllAdvised / Less Ill-Advised Modifications work with Rugged Construction
Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

Besides the "Rugged Construction" now halving the malfunction chance, it's more of a
'bonus' hullmod that brings small, 10% bonus to weapon turn rate and rate of fire
of ALL weapons - doubled if rugged is present (so, 2.5% malfunction for 20%, not bad eh)

I'm not sure it's a D-mod anymore, but either way, I'm even gonna unhide it. Make it findable.
UPDATE: made it not hidden anymore.