Issues/issue62 make dmods work with rugged construction (#65)
Issue link: 
https://github.com/RkShaRkz/Starsector-IntelliJ-Template/issues/62

This PR focuses on making vayra dmods interacting with "Rugged Construction" hullmod, and while at it - touches up a few of them by either reducing certain figures in exchange for a new defect, or just remakes it from the ground up because why not, people whined alot and I don't care that much.

Either way, "rugged construction" now halves the pain of all of these hullmods in half. And "vayra ill-advised modifications" is much less of a dmod now, and more of a ... fun thing to use when you get it I suppose. And you will get it because it's no longer hidden.


* Put in necessary methods to be used in MiscUtils

The methods are mainly regarding checking for hullmods - all 3 kinds, normally put in, built-in and S-modded,
and a convenience method that checks for Rugged Construction specifically

* Make VayraDamagedAmmo / Malfunctioning Ammo Forge work with Rugged Construction

Make it convert to FRAGMENTATION with 50% chance if "rugged" is present.

* Make VayraDamagedAutomation / Inoperative Automated Systems work with Rugged Construction

Make it apply 50% less debuff if "rugged" is present.

* Fix / improve teh DamagedAutomation rugged interaction

* Make VayraDamagedBallistics / Irregular Feed Mechanisms work with Rugged Construction

Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

* Make VayraDamagedEnvironment / Buckled Environmental Shielding work with Rugged Construction

Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

UPDATE: got rid of no-longer-used constant

* Make VayraDamagedEverything / Performance Irregularities work with Rugged Construction

Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

Also turned it up a notch, 1% malfunction seemed a bit too low now that it's fixed up and parametrizable
from luna settings. 5% (+scaling) for both weapons and engine should do fine.

But CR remained at 10% just because "you can turn it off from the settings if you don't like it, shush"

* Make VayraDamagedFighterWeapons / Malformed Fighter Weapons work with Rugged Construction

Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

Split up the recoil malus from the accuracy malus.
Decrease accuracy penalty but increase recoil penalty.

Finally, update the string in hull_mods.csv so that the recoil penalty is mentioned.

* Make VayraDamagedGunnery / Noncompliant Gunnery Core work with Rugged Construction

Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

* Extract method clamp() from DamagedAutomation to MiscUtils

* Make VayraDamagedLifeSupport / Impaired Life Support Systems work with Rugged Construction

Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

Also kinda/maybe/probably softened up the penalty by clamping it to 0, so at 200%
dmod effectiveness it will still only take up the whole 100% of ship's max crew
instead of 150% that it should have :D

* Make VayraDamagedMissiles / Flawed Missile Racks work with Rugged Construction

Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

Also, somewhat change up what the dmod does; instead of just bringing the ammo count
and RoF down, now it instead brings the ammo count and guidance down but brings RoF up!

Should be somewhat more fun to use I guess and not that much of a bore to get.

* Make VayraDamagedOptics / Distorted Optics work with Rugged Construction

Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

Also, lowered the malus from 25% to 15%, and made rugged affect both the beam range malus and
the beam waver malus.

* Make VayraDamagedShields / Destabilized Shield Generator work with Rugged Construction

Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

* Make VayraDamagedTurrets / Misshapen Turret Gyros work with Rugged Construction

Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

Changed the dmod up a bit, instead of giving 25% turnrate and health penalties,
it now gives 15% turnrate and health penalties, and a new 5% chance of weapon malfunction.

* Soften up / slightly rework VayraDamagedAmmo

I don't really care about this dmod, and people whined enough so why not.
Here. Instead of the "50% chance" being a rugged-exclusive feature, why not
make it the default one:
- 50% chance to change to fragmentation
- 25% chance to change to fragmentation if you have Rugged Construction

there. I don't wanna hear any more nonsense about this one.

* Make VayraLessIllAdvised / Less Ill-Advised Modifications work with Rugged Construction

Change up the formula so that it's effectivelly the same but expressed in terms of penalties,
and make having "rugged" chop those penalties (maluses) in half.

Besides the "Rugged Construction" now halving the malfunction chance, it's more of a
'bonus' hullmod that brings small, 10% bonus to weapon turn rate and rate of fire
of ALL weapons - doubled if rugged is present (so, 2.5% malfunction for 20%, not bad eh)

I'm not sure it's a D-mod anymore, but either way, I'm even gonna unhide it. Make it findable.
UPDATE: made it not hidden anymore.

* Soften up VayraDamagedAmmo to convert to Fragmentation with a chance

Initially, make the chance to 50%, scaled with D-mod effects, and then halved by
having Rugged Construction.

* Slight rewrite of DamagedAutomation; consolidating against strings; all fine

* Address own CR comments