Reorganized code a bit, and leftover TPCs should upgrade to cheap TPCs after s-modding
However, a few more improvements could be made to make it test-worthy.
UPDATE: actually, i'm gonna give it a spin and see what's there to see.
UPDATE2: actually - works surprisingly well, just had a issue when trying to
clear out an already-empty weapon slot and hitting a NPE when trying to check it's ID