Works as 'expected' but something is wrong
When the hullmod is installed, only one 'normal' looted TPC is installed.
When it's subsequently s-modded, all remaining slots are filled in with cheaper TPCs.

When it's un-S-modded, only one of them is replaced back with a normal TPC.

Some slight error of judgement has crept in here; need to decouple the code a bit more to get
to the bottom of it.