Fix issue reported by Extra Fluffy
Seems like there's a derelict that references a non-existing variant,
and is even mentioned/referenced in KadurMirageSystem line 290.

However, there doesn't seem to be a variant using that name in the codebase...

Extracted them to strings for now, will need to either look for that variant, or
do the cheap way out of copypasting some other variant OR changing the string to some other variant.