Try to tackle issue69 which might be due to lacking thread safety
Even though everything in VayraPersonBountyIntel is instance-based,
none of this thread-safety should have needed to be put in in the first place.

But, the only explanation to why starsector API is hitting a nullpointer on something,
is that either the skills themselves were removed by some other mod, or that a race-condition
is happening and stepping all over the toes of previous execution, clearing up it's members
(which it most likely isnt)

But eiter way, more thread-safety wont hurt.