Improve VayraLessIllAdvised interaction with "rugged"
Since it seems that "rugged" is already implicitly a part of baseEffect, we do **NOT**
need to check for it manually and reduce the effect to half if it's present.

This is something that needs to be done for every other dmod as well, but in case of
VayraLessIllAdvised - we will also not just double the 'bonuses' - because that matches
the other non-rugged ship DMod values, but rather quadruple it, so that it effectivelly
ends up with double the bonuses and half the penalties from the "less ill advised" dmod.

EDIT: applied the quadrupling to turn rate bonus as well.