Consolidate (fix) how other DMods work with "rugged" construction hullmod
As mentioned before, "rugged" is already implicitly included in the "base effect"
or rather the Stats.DMOD_EFFECT_MULT - normally 1.0, with "rugged" it's 0.5

So the whole "let's halve the effects if we have rugged" was effectivelly
quartering them.

So now with all of that removed, more cleanup was done because suddenly the whole
ShipVariantAPI parameter (and argument) were unnecessary everywhere, so they were
removed as well.

This completely wraps up the final Vayra DMod reworking for a good time,
since it not only works as expected with rugged construction, but also allows
full parametrization via luna settings. To each their own.