Issues/issue77 parametrize vayra dmods values (#78)
Issue link:
https://github.com/RkShaRkz/Starsector-IntelliJ-Template/issues/77

This PR attempts to resolve issue (feature request) #77 to make DMod numbers parametrizable since the turning them on/off for yourself or the enemy wasn't fair either because people would also like something in between.

In an attempt to not be unfair to others while being fair to some weeklings (or vice-versa) exposing all of the DMods' parameters to the users to customize seems like a move that would bring value to everyone.

* Make VayraDamagedAmmo effect values parametrized through LunaSettings

* Make VayraDamagedAutomation effects' values parametrized through LunaSettings

* Make VayraDamagedBallistics effects' values parametrized through LunaSettings

* Make VayraDamagedEnvironment effects' values parametrized through LunaSettings

* Make VayraDamagedEverything effects' values parametrized through LunaSettings

EDIT: throw in case it's neither of the two.

* Make VayraDamagedFighterWeapons effects' values parametrized through LunaSettings

* Make VayraDamagedGunnery effects' values parametrized through LunaSettings

* Consolidate DamagedBallistics in LunaSettings.csv

It's default value was well outside the allowed range.

* Consolidate DamagedAutomation in LunaSettings nitpicks

The 'double' number didn't really have a decimal so...

* Make VayraDamagedLifeSupport effects' values parametrized through LunaSettings

* Make VayraDamagedMissiles effects' values parametrized through LunaSettings

EDIT: fix the lunasettngs.csv field names, they had some leftover from copypasting...

* Make VayraDamagedOptics effects' values parametrized through LunaSettings

* Make VayraDamagedShields effects' values parametrized through LunaSettings

* Make VayraDamagedTurrets effects' values parametrized through LunaSettings

* Latest JAR

* Improve VayraLessIllAdvised interaction with "rugged"

Since it seems that "rugged" is already implicitly a part of baseEffect, we do **NOT**
need to check for it manually and reduce the effect to half if it's present.

This is something that needs to be done for every other dmod as well, but in case of
VayraLessIllAdvised - we will also not just double the 'bonuses' - because that matches
the other non-rugged ship DMod values, but rather quadruple it, so that it effectivelly
ends up with double the bonuses and half the penalties from the "less ill advised" dmod.

EDIT: applied the quadrupling to turn rate bonus as well.

* Make VayraLessIllAdvised effects' values parametrized through LunaSettings

* Consolidate (fix) how other DMods work with "rugged" construction hullmod

As mentioned before, "rugged" is already implicitly included in the "base effect"
or rather the Stats.DMOD_EFFECT_MULT - normally 1.0, with "rugged" it's 0.5

So the whole "let's halve the effects if we have rugged" was effectivelly
quartering them.

So now with all of that removed, more cleanup was done because suddenly the whole
ShipVariantAPI parameter (and argument) were unnecessary everywhere, so they were
removed as well.

This completely wraps up the final Vayra DMod reworking for a good time,
since it not only works as expected with rugged construction, but also allows
full parametrization via luna settings. To each their own.

* Latest JAR