Issue #76 - check vayra factions console command - initial checkpoint commit
Issue link:
https://github.com/RkShaRkz/Starsector-IntelliJ-Template/issues/76

This commit adds the initial barebones console command to check the status of the
following factions and their target markets, and then decides whether the factions
are enabled, have spawned and are active - or not.

Factions that are checked:
communist_clouds
science_fuckers
warhawk_republic
almighty_dollar
ashen_keepers
kadur_remnant

For now, the command first tries to get the market (planet), and if that doesn't exist
it tries getting the market from the station and stringifies some shared details about that market.

That part needs to be unified and refactored but for now - all 6 of these factions report
that they are enabled and stringify some details about their future home.

So before restructuring code, it's a good time to save this work in case i mess it up.