Slight refactor to VayraColonialManager and fix the issue I left this morning
Issue being - the game was crashing during load. The reason for that being - guess what,
there's no VayraColonialManager instance to touch in mainmenu because it's an OnEveryFrameScript.

tl;dr - the IntervalUtils had to be moved *outside* of the OnEveryFrameScript into a class we can touch,
and just let the OnEveryFrameScript use those instances instead of their own.

The VayraColonialExpeditionIntel just got a small @NonNull annotation to indicate that the Faction parameter
must never be null.