Rework the whole VayraColonialExpeditionIntel internals and fix some logs in VayraColonialManager
the other two files were just affected by the refactoring.

Anyways, the real problem was - VayraColonialExpeditionIntel was calling into a 'super' constructor,
passing a non-null faction.

However, the super (RaidIntel) constructor was properly using getFaction() to get the faction passed
to the constructor instead of touching the field by it's member reference.

Which led to the convoluted problem:
- VayraColonialExpeditionIntel was calling super constructor
- VayraColonialExpeditionIntel was using a different field for the same thing that the 'super' is using 'faction'
- VayraColonialExpeditionIntel was overriding the getFaction() method to return *it's* field.

So somewhere during the super(...) call, the getFaction() is getting called, but because the "alternative faction" is set
after the super(...) call, it was always null. Thus, the VayraColonialExpeditionIntel never managed to finish construction
because it was throwing a NPE before any of it's fields got set.

Of course, this was fixed by deleting the override getFaction() and using the superclass's 'faction' for everything it was
using it's alternate `colonyFaction` field.

Furthermore, it's "internal" RaidDelegate (i'm gonna call it "internal" even though the whole VayraColonialExpeditionIntel
implements RaidDelegate) has been merged with my attempt to log the progress of the colonial expedition.

As far as logs go, some markets were being logged in a unneccessarily long manner in VayraColonialManager, so we're gonna use
a shorter, more informative way of logging them.

These two tiny things still need to be re-verified but besides that - colonial expeditions work fine and show up in the Intel screen fine.