Improve the nullguarding of PersonBountyManager.getInstance() call
It just occured to me that nullguarding the call itself might not fix the problem
because the shitty `PersonBountyManager.getInstance()` might return null between two
calls, which shouldn't be happening but this is Alex's Starsector API after all.

So, grab the instance into a variable and then nullguard that thing. That will cause it
to not mutate between two different getInstance() calls and should be safer.