Give up on Field Repairs skill
The knobs and everything in luna settings remains, but they do nothing.
Even when I strip the fields of their final modifier, and manage changing them
(or not) - nothing effectivelly changes in the skill.

Maybe due to constant inlining at compile-time? Who knows.

Why not tackle other skills instead of being stuck on this one.